change to size_t in basevector set



See merge request !95

+++ linalg/basevector.cpp
@@ -101,7 +101,7 @@ namespace ngla
     t.AddFlops (fv.Size());
 
     ParallelFor ( fv.Range(),
-                  [fv,scal] (int i) { fv(i) = scal; });
+                  [fv,scal] (size_t i) { fv(i) = scal; });
     
     return *this; 
   }
